+++ packages/akutil-common/CHANGELOG.md
@@ -1,5 +1,10 @@
-<a name="0.0.3"></a>
-## 0.0.3 (2016-07-04)
+<a name="0.1.0"></a>
+# [0.1.0](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/compare/0.1.0-semver-tag-for-akutil-common...v0.1.0) (2016-07-05)
+
+
+
+<a name="0.1.0-semver-tag-for-akutil-common"></a>
+# 0.1.0-semver-tag-for-akutil-common (2016-07-05)