+++ packages/eslint-config-ak-base/CHANGELOG.md
@@ -1,10 +1,25 @@
-<a name="1.0.2"></a>
-## [1.0.2](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/compare/1.0.1-semver-tag-for-eslint-config-ak-base...v1.0.2) (2016-07-04)
+<a name="1.1.0"></a>
+# [1.1.0](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/compare/1.1.0-semver-tag-for-eslint-config-ak-base...v1.1.0) (2016-07-05)
 
 
 
-<a name="1.0.1-semver-tag-for-eslint-config-ak-base"></a>
-## 1.0.1-semver-tag-for-eslint-config-ak-base (2016-06-27)
+<a name="1.1.0-semver-tag-for-eslint-config-ak-base"></a>
+# [1.1.0-semver-tag-for-eslint-config-ak-base](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/compare/1.0.3-semver-tag-for-eslint-config-ak-base...1.1.0-semver-tag-for-eslint-config-ak-base) (2016-07-05)
+
+
+
+<a name="1.0.3-semver-tag-for-eslint-config-ak-base"></a>
+## [1.0.3-semver-tag-for-eslint-config-ak-base](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/compare/bd59ebf...1.0.3-semver-tag-for-eslint-config-ak-base) (2016-07-05)
+
+
+### Bug Fixes
+
+* empty ([c7fc0fc](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/commits/c7fc0fc))
+
+
+### Features
+
+* **lint:** Add support for JSX in the eslint-config-ak-base and add docs for it. ([bd59ebf](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/commits/bd59ebf))