+++ packages/eslint-config-ak-cucumber/CHANGELOG.md
@@ -1,5 +1,5 @@
 <a name="1.0.2"></a>
-## [1.0.2](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/compare/1.0.1-semver-tag-for-eslint-config-ak-cucumber...v1.0.2) (2016-07-04)
+## [1.0.2](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/compare/1.0.1-semver-tag-for-eslint-config-ak-cucumber...v1.0.2) (2016-07-05)